refactor(database): reorganize `database/file.c`
